- warning suppression; - msectime_type is not uint64_t.

+++ src/scag/util/HexDump.h
@@ -11,16 +11,16 @@ class HexDump
 {
 public:
     // requested size of the buffer
-    unsigned hexdumpsize( unsigned insize ) const { return insize*3; }
-    char* hexdump( char* outbuf, const char* inbuf, unsigned insize );
-    unsigned strdumpsize( unsigned insize ) const { return insize; }
-    char* strdump( char* outbuf, const char* inbuf, unsigned insize );
+    size_t hexdumpsize( size_t insize ) const { return insize*3; }
+    char* hexdump( char* outbuf, const char* inbuf, size_t insize );
+    size_t strdumpsize( size_t insize ) const { return insize; }
+    char* strdump( char* outbuf, const char* inbuf, size_t insize );
 
     // NOTE: w/o termination null-byte
     char* addstr( char* outbuf, const char* cstring );
 
-    void hexdump( std::string& out, const char* inbuf, unsigned insize );
-    void strdump( std::string& out, const char* inbuf, unsigned insize );
+    void hexdump( std::string& out, const char* inbuf, size_t insize );
+    void strdump( std::string& out, const char* inbuf, size_t insize );
 
 private:
     static std::auto_ptr<char> digits_;
